requirements.txt
langchain
langchain-openai
langchain-community
python-dotenv
faiss-cpu
tiktoken
requests
fastapi
uvicorn
sqlalchemy